Extract Slider Label to own component
